copy settings button already big enough for its own module
